improvement: rename some things
