+++ index-files/index-1.html
@@ -52,12 +52,12 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">C</a>&nbsp;<a href="index-2.html">P</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:C">C</h2>
 <dl class="index">
-<dt><a href="../com.tcs.edu.printerr.ConsolePrinterr.html" class="type-name-link" title="class in Unnamed Package">com.tcs.edu.printerr.ConsolePrinterr</a> - Class in <a href="../package-summary.html">Unnamed Package</a></dt>
+<dt><a href="../com.tcs.edu.printer.ConsolePrinterr.html" class="type-name-link" title="class in Unnamed Package">com.tcs.edu.printer.ConsolePrinterr</a> - Class in <a href="../package-summary.html">Unnamed Package</a></dt>
 <dd>
-<div class="block">Класс com.tcs.edu.printerr.ConsolePrinterr
+<div class="block">Класс com.tcs.edu.printer.ConsolePrinterr
     Включает в себя методы связанные с выводом сообщений в консоль</div>
 </dd>
-<dt><a href="../com.tcs.edu.printerr.ConsolePrinterr.html#%3Cinit%3E()" class="member-name-link">com.tcs.edu.printerr.ConsolePrinter()</a> - Constructor for class <a href="../com.tcs.edu.printerr.ConsolePrinterr.html" title="class in Unnamed Package">com.tcs.edu.printerr.ConsolePrinterr</a></dt>
+<dt><a href="../com.tcs.edu.printer.ConsolePrinterr.html#%3Cinit%3E()" class="member-name-link">com.tcs.edu.printer.ConsolePrinter()</a> - Constructor for class <a href="../com.tcs.edu.printer.ConsolePrinterr.html" title="class in Unnamed Package">com.tcs.edu.printer.ConsolePrinterr</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">C</a>&nbsp;<a href="index-2.html">P</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
