+++ index-files/index-2.html
@@ -52,7 +52,7 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">C</a>&nbsp;<a href="index-2.html">P</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:P">P</h2>
 <dl class="index">
-<dt><a href="../com.tcs.edu.printerr.ConsolePrinterr.html#print(java.lang.String)" class="member-name-link">print(String)</a> - Static method in class <a href="../com.tcs.edu.printerr.ConsolePrinterr.html" title="class in Unnamed Package">com.tcs.edu.printerr.ConsolePrinterr</a></dt>
+<dt><a href="../com.tcs.edu.printer.ConsolePrinterr.html#print(java.lang.String)" class="member-name-link">print(String)</a> - Static method in class <a href="../com.tcs.edu.printer.ConsolePrinterr.html" title="class in Unnamed Package">com.tcs.edu.printer.ConsolePrinterr</a></dt>
 <dd>
 <div class="block">Метод print
  Предназначен для вывода сообщений в консоль</div>
