+++ src.done/main/java/com/tcs/edu/Application.java
@@ -1,7 +1,7 @@
 package com.tcs.edu;
 
-import com.tcs.edu.decoratorr.TimestampMessageDecorator;
-import com.tcs.edu.printerr.ConsolePrinter;
+import com.tcs.edu.decorator.TimestampMessageDecorator;
+import com.tcs.edu.printer.ConsolePrinter;
 
 /**
  * Main class: entry point
