+++ src.done/main/java/com/tcs/edu/decorator/TimestampMessageDecorator.java
@@ -1,4 +1,4 @@
-package com.tcs.edu.decoratorr;
+package com.tcs.edu.decorator;
 
 import java.time.Instant;
